+++ test/specs/languages/test-css.el
@@ -11,6 +11,24 @@
       (flycheck-buttercup-should-syntax-check
        "language/css/syntax-error.css" 'css-mode
        '(4 5 error "Unknown word font-size (CssSyntaxError)"
-           :id "CssSyntaxError" :checker css-stylelint)))))
+           :id "CssSyntaxError" :checker css-stylelint))))
+
+  (describe "the stylelint checker command"
+    (it "always requests JSON output"
+      (let ((flycheck-stylelintrc nil)
+            (flycheck-stylelint-args nil))
+        (expect (flycheck-checker-substituted-arguments 'css-stylelint)
+                :to-contain "--formatter")
+        (expect (flycheck-checker-substituted-arguments 'css-stylelint)
+                :to-contain "json")))
+
+    (it "appends flycheck-stylelint-args after the JSON formatter"
+      (let ((flycheck-stylelintrc nil)
+            (flycheck-stylelint-args '("--custom-syntax" "postcss-scss")))
+        (let ((args (flycheck-checker-substituted-arguments 'scss-stylelint)))
+          (expect args :to-contain "--custom-syntax")
+          (expect args :to-contain "postcss-scss")
+          ;; the mandatory JSON formatter is still there
+          (expect args :to-contain "json"))))))
 
 ;;; test-css.el ends here